+++ tket/tests/tkettestsfiles.cmake
@@ -56,6 +56,7 @@ set(TEST_SOURCES
     ${TKET_TESTS_DIR}/Circuit/test_Multiplexor.cpp
     ${TKET_TESTS_DIR}/test_CliffTableau.cpp
     ${TKET_TESTS_DIR}/test_UnitaryTableau.cpp
+    ${TKET_TESTS_DIR}/test_ChoiMixTableau.cpp
     ${TKET_TESTS_DIR}/test_PhasePolynomials.cpp
     ${TKET_TESTS_DIR}/test_PauliGraph.cpp
     ${TKET_TESTS_DIR}/test_Architectures.cpp