+++ pytket/docs/changelog.rst
@@ -11,12 +11,36 @@ Major new features:
 
 Minor new features:
 
+* Circuit methods ``qubit_readout`` and ``qubit_to_bit_map`` now ignore barriers.
+
+1.10.0 (December 2022)
+----------------------
+
+Minor new features:
+
+* Add support for PhasedX gates in Pauli graph synthesis.
+
+Fixes:
+
+* Handle 0-qubit operations in connectivity check.
+* Fix handling of Tdg, CY, ZZMax and Clifford-angle YYPhase gates in Pauli
+  graph synthesis.
+* Disallow conversion to QASM of operations conditioned on strict subregisters
+  larger than one bit, or reordered registers.
+
+1.9.1 (December 2022)
+---------------------
+
+Minor new features:
+
 * New ``view_browser`` function for opening a browser with circuit render.
 
 Fixes:
 
 * Warn rather than abort when significant rounding errors are detected in
   TK2-to-CX rebase.
+* Fix incorrect QASM output for ``OpType.CopyBits``.
+* Fix incorrect QASM read in ``OpType.ZZPhase``.
 
 1.9.0 (November 2022)
 ---------------------